refactor: enforce filename casing consistency
